Remove (almost) non-functional and deprecated RMS env settings
Backwards compatibility probably means that it exists in user
configurations, but Ert will not do anything if a RMS forward model is
configured with an unused option, like:

  FORWARD_MODEL RMS(..., <RMS_SOMETHING_UNRECOGNIZED>=FOOBAR)

so it should be safe to remove altogether.

When present, these two variables yield two warnings emitted pr.
realization due to the variables not being defined.